ACTION Orbital Debris Concept
ACTION: Advanced Capture and Transport Integrated deOrbiting Nanosat

6U CubeSat capable of removing three debris objects per mission from LEO, called Advanced Capture and Transport Integrated deOrbiting Nanosat (ACTION). The design is flexible, cheap, and scalable. Launch costs for nanosats are far less than for dedicated large ADR missions, which gives ACTION an order of magnitude cost advantage. See the table [in PDF] for a comparison between ACTION and large ADR concepts, at the recommended pace of five large objects per year (Liou, 2011, p. 59). Each ACTION unit contains three Capture, Containment, De-spin and Drag (CCDD) modules and three small Thruster-Deorbiting Kits (TDKs). The CCDD module expands after being deployed and uses a drawstring closure to contain debris. A 20-meter electromagnetic tether is then deployed, along with a counterweight (the “yoyo”) in order to stabilize spinning debris after capture by transferring angular momentum to the yoyo as the tether deploys. 

The CCDD module stows with 1/5 of its deployed diameter and a 35x reduction in surface area (over 48x when counting the area of the deployed tether), enabling three of the modules to be stowed and deployed from a volume of 2 L. Folding is accomplished by first collapsing the structure, and then folding the frame into 5 overlapping loops. A mathematical analysis of this folding technique is provided in Mouthuy, et. al., 2012 as shown in Fig. 5. Our team has concluded that this technique has never been applied to use in space. The CCDD module structure consists of two or three frame pieces, a spring-like frame piece, and a carbon-Kevlar-reinforced mylar “bag” wrapper that acts as a tensioned skin once deployed. 

After deployment, the CCDD module is stabilized by a basic Reaction Control System (RCS) and targets the debris using LIDAR sensing or a computer vision system. The  targeting computations for autonomous debris capture take place inside the ACTION unit using sensors mounted inside the deployed CCDD module. The drawstring closure and tether deployment should be activated by a simple analog circuit or mechanical device that is armed before the encounter with the debris object; detecting the encounter via an impact sensor, vibration sensor, or other reliable analog mechanism. After containment of the debris object, a TDK can be deployed and maneuver to the end of the tether, docking to the yoyo using an electromagnet on the front of the TDK to attract the yoyo.